requirements.txt
# Core dependencies for SOP processing
openai>=1.0.0
PyPDF2>=3.0.0
python-docx>=1.0.0

# Existing framework dependencies (if needed)
# torch>=2.0.0
# numpy>=1.24.0
# pyyaml>=6.0

